requirements.txt
absl-py==0.5.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
atari-py==0.1.1
atomicwrites==1.2.1
attrs==18.1.0
Automat==0.3.0
autovizwidget==0.12.5
backcall==0.1.0
bleach==2.1.4
bokeh==0.13.0
boto3==1.7.82
botocore==1.10.82
certifi==2018.8.13
cffi==1.11.5
characteristic==14.3.0
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.3.1
cryptography-vectors==2.3.1
cycler==0.10.0
Cython==0.28.4
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
environment-kernels==1.1.1
flatbuffers==2015.12.22.1
funcsigs==1.0.2
future==0.16.0
gast==0.2.0
grpcio==1.15.0
gym==0.10.5
h5py==2.8.0
hdijupyterutils==0.12.5
html5lib==1.0.1
hyperlink==17.3.1
idna==2.7
incremental==17.5.0
ipykernel==4.8.2
ipyparallel==6.2.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
jedi==0.12.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
keras-rl==0.4.2
kiwisolver==1.0.1
lz4==2.1.0
Markdown==3.0
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
more-itertools==4.3.0
nb-conda==2.2.1
nb-conda-kernels==2.1.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.5
olefile==0.45.1
onnx==1.2.1
opencv-python==3.4.3.18
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
plotly==3.1.1
pluggy==0.7.1
prometheus-client==0.3.0
prompt-toolkit==1.0.15
protobuf==3.6.1
protobuf3-to-dict==0.1.5
psutil==5.4.7
psycopg2==2.7.5
ptyprocess==0.6.0
py==1.6.0
py4j==0.10.4
pyasn1==0.4.4
pyasn1-modules==0.2.1
pycparser==2.18
pygal==2.4.0
pyglet==1.3.2
Pygments==2.2.0
PyHamcrest==1.9.0
pykerberos==1.1.14
PyOpenGL==3.1.0
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
pyspark==2.2.1
pytest==3.8.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
ray==0.5.2
redis==2.10.6
requests==2.19.1
requests-kerberos==0.12.0
retrying==1.3.3
s3fs==0.1.5
s3transfer==0.1.13
sagemaker==1.9.0
sagemaker-pyspark==1.1.3
scikit-learn==0.19.2
scipy==1.1.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
sparkmagic==0.12.5
SQLAlchemy==1.2.11
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
torch==0.4.1
torchvision==0.2.1
tornado==5.1
traitlets==4.3.2
Twisted==18.7.0
typing==3.6.4
typing-extensions==3.6.5
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.14.1
widgetsnbextension==3.4.0
zope.interface==4.5.0
